Move EdgeType out of graph module
