+++ resotolib/test/test_graph.py
@@ -1,7 +1,8 @@
 import pytest
-from resotolib.graph import Graph, GraphContainer, EdgeType, GraphExportIterator
-from resotolib.baseresources import BaseResource, GraphRoot
 import resotolib.logging as logging
+from resotolib.graph import Graph, GraphContainer, GraphExportIterator
+from resotolib.graph.edges import EdgeType
+from resotolib.baseresources import BaseResource
 from dataclasses import dataclass
 from typing import ClassVar
 from sys import getrefcount
